Converting CPP server to C99 Server

+++ Cpp/Enter the Deep Server/DeepEngine-Archetype.cpp
@@ -4,7 +4,7 @@ namespace DeepEngine
 {
 	ComponentList Archetype::GetComponent(ECSHandle SingleType)
 	{
-		for (int i = 0; i < Type.size(); ++i)
+		for (size_t i = 0; i < Type.size(); ++i)
 		{
 			if (Type[i] == SingleType) return Components[i];
 		}
